requirements.txt
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
anthropic==0.40.0
python-dotenv==1.0.1
schedule==1.2.2
rich==13.9.4
twilio==9.3.7
